Split GameUI.tsx into HomeScreen, lobby and tournament modules
Pure move, no behaviour change. GameUI.tsx was 2992 lines holding five
screens; it is now the ~30-line router between them:

- components/ui/HomeScreen.tsx      — landing screen, public-lobby/live-game lists
- components/lobby/LobbyOverlay.tsx — sealed/draft/premade lobby + premade deck picker
- components/tournament/TournamentOverlay.tsx — standings, deck viewer, save-deck dialog
- components/tournament/FreeForAllOverlay.tsx — FFA pod standings
- components/ui/FullscreenButton.tsx — shared by the landing and lobby screens

Phase 0 of backlog/menu-lobby-restructure-and-help.md — enabler for the
landing restructure and the unified lobby.
