+++ web-client/src/components/ui/FullscreenButton.tsx
@@ -1,0 +1,39 @@
+import { useState, useEffect } from 'react'
+import styles from './GameUI.module.css'
+
+/**
+ * Fullscreen toggle button. Pinned to the corner of the landing and lobby screens.
+ */
+export function FullscreenButton() {
+  const [isFullscreen, setIsFullscreen] = useState(false)
+
+  useEffect(() => {
+    const handleFullscreenChange = () => {
+      setIsFullscreen(!!document.fullscreenElement)
+    }
+    document.addEventListener('fullscreenchange', handleFullscreenChange)
+    return () => document.removeEventListener('fullscreenchange', handleFullscreenChange)
+  }, [])
+
+  const toggleFullscreen = async () => {
+    try {
+      if (!document.fullscreenElement) {
+        await document.documentElement.requestFullscreen()
+      } else {
+        await document.exitFullscreen()
+      }
+    } catch (err) {
+      console.error('Fullscreen error:', err)
+    }
+  }
+
+  return (
+    <button
+      onClick={toggleFullscreen}
+      className={styles.fullscreenButton}
+      title={isFullscreen ? 'Exit fullscreen (Esc)' : 'Enter fullscreen'}
+    >
+      {isFullscreen ? '⛶ Exit' : '⛶ Fullscreen'}
+    </button>
+  )
+}
